% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.ncpt}
\alias{ci.ncpt}
\title{Compute confidence interval for noncentral t}
\usage{
ci.ncpt(ncp, k, n, conf.level = 0.95)
}
\arguments{
\item{ncp}{noncentrality parameter}

\item{k}{number of groups.}

\item{n}{cell size (for each group)}

\item{conf.level}{width of the confidence interval, default=.95}
}
\value{
confidence intervals.
}
\description{
Return the confidence intervals computed by scaling the confidence intervals of the noncentrality parameter.
}
\details{
Compute the confidence interval based on the Steiger and Fouladi (1997) method for the noncentrality parameter of t
}
\references{
Steiger, J. H., & Fouladi, R. T. (1997). Noncentrality interval estimation and the evaluation of statistical models. In L. L. E. Harlow, S. A. Mulaik, & J. H. Steiger (Eds.), What if there were no significance tests?: Classic edition. Lawrence Erlbaum Associates Publishers.
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
