% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{contr.custom}
\alias{contr.custom}
\title{build a custom contrasts and k-2 orthogonal ones}
\usage{
contr.custom(cont)
}
\arguments{
\item{cont}{Character string indicating a set of colors.}
}
\value{
a data.frame with k-1 contrast codes, with k=length(cont).
}
\description{
Creates a set of contrast codes that are orthogonal to a given one with the aim of
testing the given contrast in a linear model
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
