% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variances.R
\name{d.contr.eta2}
\alias{d.contr.eta2}
\title{Compute contrast d from (partial) eta-squared}
\usage{
d.contr.eta2(cont, eta2, scale = "g")
}
\arguments{
\item{cont}{vector of contrast weights.}

\item{eta2}{eta-squared coefficient for given contrast.}

\item{scale}{method to scale d.}
}
\value{
eta-squared.
 @details
     The parameter \code{scale} controls the method used to scale the effect size d.
     \enumerate{
    \item    \code{scale="g"} assumes scaling by dividing 2*d by the sum of absolute coefficients
    \item    \code{scale="q"} assumes scaling by dividing d by the square-root of the sum of squares of the coefficients
    \item    \code{numeric} any constant that multiplies the unscaled d to obtain the scaled d
   }
}
\description{
Compute contrast d from (partial) eta-squared
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
