% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{d.contr}
\alias{d.contr}
\title{Cohen's d for contrasts}
\usage{
d.contr(cont, means = NULL, sd = NULL, y = NULL, x = NULL,
  scale = "g")
}
\arguments{
\item{cont}{contrast codes as a numeric vector}

\item{means}{pattern of means as numeric vector of the same length of \code{cont}}

\item{sd}{pooled standard deviation}

\item{y}{dependent variable}

\item{x}{independent variable with k-groups, with k being length(cont)}

\item{scale}{scaling method of the d index}
}
\value{
t-test for the contrast.
}
\description{
Compute Cohen's d for contrasts with different scaling functions
}
\details{
data to compute the d index can be either actual data or a vector of means and a standard deviation
         actual data are specified with the \code{y} and \code{x} parameters, means and standard deviation
         with the \code{means} , and \code{sd} parameters. The parameter \code{means} and \code{sd} are prevalent, thus
         if \code{means},\code{y} and \code{x} are provided, but not \code{sd}, y and x are used to computed the pooled standard deviation.
         If \code{means} and \code{sd} are provided, \code{y} and \code{x} are ignored.

     The parameter \code{scale} controls the method used to scale the effect size d.
     \enumerate{
    \item    \code{scale="g"} assumes scaling by dividing 2*d by the sum of absolute coefficients
    \item    \code{scale="q"} assumes scaling by dividing d by the square-root of the sum of squares of the coefficients
    \item    \code{numeric} any constant that multiplies the unscaled d to obtain the scaled d
   }
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
