% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variances.R
\name{f.contr.d}
\alias{f.contr.d}
\title{Compute contrast f from d}
\usage{
f.contr.d(cont, d, scale = "g")
}
\arguments{
\item{cont}{vector of contrast weights.}

\item{d}{eta-squared coefficient for given contrast.}

\item{scale}{method to scale d.}
}
\value{
f.
}
\description{
Compute contrast f from d
}
\details{
The parameter \code{scale} controls the method used to scale the effect size d.
     \enumerate{
    \item    \code{scale="g"} assumes scaling by dividing 2*d by the sum of absolute coefficients
    \item    \code{scale="z"} assumes scaling by dividing d by the square-root of the sum of squares of the coefficients
    \item    \code{numeric} any constant that multiplies the unscaled d to obtain the scaled d
   }
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
