% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{one.sample}
\alias{one.sample}
\title{Return a sample with the specified pattern of means and standard deviation}
\usage{
one.sample(cont, means, sd = 1, n = 100)
}
\arguments{
\item{cont}{contrast codes as a numeric vector}

\item{means}{pattern of means as numeric vector of the same length of \code{cont}}

\item{sd}{pooled standard deviation}

\item{n}{per cell n}
}
\value{
a data.frame with all the simulated results of class simulations
}
\description{
Simulates samples based on normal distribution and test the contrast for each sample, providing
power information
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
