% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{power.contrast.f}
\alias{power.contrast.f}
\title{power function for contrasts}
\usage{
power.contrast.f(cont = NULL, d = NULL, n = NULL, power = NULL,
  scale = "g", sig.level = 0.05)
}
\arguments{
\item{cont}{contrast codes as a numeric vector}

\item{d}{effect size d}

\item{n}{number of observations per cell}

\item{power}{power of test (1 minus Type II error probability)}

\item{sig.level}{significance level (Type I error probability)}

\item{method}{used to scale the d index}
}
\value{
Object of class '"power.htest"', a list of the arguments (including the computed one) augmented with 'method' and 'note' elements.
}
\description{
computes power parameters for F-test associated with a custom contrast
}
\details{
Exactly one of the parameters 'd','n','power'
     and 'sig.level' must be passed as NULL, and that parameter is determined from the others.
     The parameter \code{scale} controls the method used to scale the effect size d.
     \enumerate{
    \item    \code{scale="g"} assumes scaling by dividing 2*d by the sum of absolute coefficients
    \item    \code{scale="q"} assumes scaling by dividing d by the square-root of the sum of squares of the coefficients
    \item    \code{numeric} any constant that multiplies the unscaled d to obtain the scaled d
   }
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
