% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{print}
\alias{print}
\title{Nice printing for confidence intervals}
\usage{
print(obj)
}
\arguments{
\item{obj}{an object of class "conf.intervals"}
}
\description{
Nice printing for confidence intervals
}
\seealso{
\code{\link[cpower]{ci.contr}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
