% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{summary}
\alias{summary}
\title{Compute simulated power expected results}
\usage{
summary(obj)
}
\arguments{
\item{obj}{a data.frame output of \code{\link[cpower]{power.contr.simulate}}}
}
\description{
Compute simulated power from power.contr.simulate() results
}
\examples{
cont<-c(-3,-1,1,3)
means<-c(10,12,10,12)
pwr<-power.contr.simulate(10,cont,10,means,sd=1)
summary(pwr)
}
\seealso{
\code{\link[cpower]{power.contr.simulate}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
