% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{summary.simulations}
\alias{summary.simulations}
\title{Compute simulated power expected results}
\usage{
\method{summary}{simulations}(obj)
}
\arguments{
\item{obj}{a data.frame outout of \code{\link[cpower]{power.contr.simulate}}}
}
\description{
Compute simulated power from power.contr.simulate() results
}
\seealso{
\code{\link[cpower]{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
