% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test.contr}
\alias{test.contr}
\title{t-test for a custom contrast}
\usage{
test.contr(data, yname, xname, cont, debug = FALSE)
}
\arguments{
\item{data}{data.frame containing the variables to be analysed}

\item{yname}{name (string) of the dependent variable}

\item{xname}{name (string) of the independent variable with k-groups, with k being length(cont)}

\item{cont}{contrast codes as a numeric vector with length=k}

\item{debug}{if TRUE, prints out the full model results. If FALSE (default) prints only the contrast results}
}
\value{
t-test for the contrast.
}
\description{
t-test for custom contrast
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
