% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{contr.custom}
\alias{contr.custom}
\title{Build a custom contrast and k-2 orthogonal ones}
\usage{
contr.custom(cont)
}
\arguments{
\item{cont}{vector of contrast weight (numeric).}
}
\value{
a data.frame with k-1 contrast codes, with k=length(cont). It can be used to assing contrast weights using \code{\link{contrasts}}
}
\description{
Creates a set of contrast codes that are orthogonal to a given one with the aim of
testing the given contrast in a linear model
}
\examples{
### sim some data
cont<-c(-3,-1,1,3)
means<-c(10,12,10,12)
y<-rep(means,1000)+rnorm(4000,0,2)
x<-factor(rep(1:4,1000))
### assign contrast weights
contrasts(x)<-contr.custom(cont)
contrasts(x)
cor(contr.custom(cont))
summary(lm(y~x))
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
