% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variances.R
\name{eta2.contr.d}
\alias{eta2.contr.d}
\title{Compute (partial) eta-squared for contrast d}
\usage{
eta2.contr.d(cont, d, scale = "g")
}
\arguments{
\item{cont}{vector of contrast weights.}

\item{d}{d coefficient for given contrast.}

\item{scale}{method to scale d.}
}
\value{
eta-squared.
}
\description{
Return the proportion of variance explained by the contrast over the variance not explained by other effects.
The result is the partial eta-squared unless, there's only one possible contrast in the model, the is the eta-squared
}
\details{
The parameter \code{scale} controls the method used to scale the effect size d.
     \enumerate{
    \item    \code{scale="g"} assumes scaling by dividing 2*d by the sum of absolute coefficients
    \item    \code{scale="q"} assumes scaling by dividing d by the square-root of the sum of squares of the coefficients
    \item    \code{numeric} any constant that multiplies the unscaled d to obtain the scaled d
   }
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
