% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variances.R
\name{eta2.contr}
\alias{eta2.contr}
\title{Compute (partial) eta-squared for contrast}
\usage{
eta2.contr(cont, xname, model)
}
\arguments{
\item{cont}{vector of contrast weights.}

\item{xname}{Character string indicating the name of the factor defining groups.}
}
\value{
eta-squared.
}
\description{
Return the proportion of variance explained by the contrast over the variance not explained by other effects.
The result is the partial eta-squared unless, there's only one possible contrast in the model, the is the eta-squared
}
\details{
Compute the eta-squared for a contrast based on a model already estimated with the effect of
"xname" af a factor.
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
