% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{power.contr.simulate}
\alias{power.contr.simulate}
\title{Compute power based on simulations}
\usage{
power.contr.simulate(rep, cont, n, means, sd = 1)
}
\arguments{
\item{cont}{contrast codes as a numeric vector}

\item{n}{per cell n}

\item{means}{pattern of means as numeric vector of the same length of \code{cont}}

\item{sd}{pooled standard deviation}
}
\value{
a data.frame with all the simulated results of class simulations
}
\description{
Simulates samples based on normal distribution and test the contrast for each sample, providing
power information
}
\seealso{
\code{\link{cpower}}
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
