% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test.contr}
\alias{test.contr}
\title{t-test for a custom contrast}
\usage{
test.contr(cont, d = NULL, n = NULL, scale = NULL, y = NULL, x = NULL)
}
\arguments{
\item{cont}{contrast codes as a numeric vector with length=k}

\item{d}{effect size index to test}

\item{n}{total sample size required if d is provided}

\item{scale}{scaling method used to compute d. Required if d is provided}

\item{data}{data.frame containing the variables to be analysed}

\item{yname}{name (string) of the dependent variable}

\item{xname}{name (string) of the independent variable with k-groups, with k being length(cont)}
}
\value{
t-test for the contrast. The \code{Estimate} is the contrast value estimate from the data and the \code{Std. Error} is expressed in the same
        scale of the estimate. It is assumed that the cells have the same size.
}
\description{
Perform a t-test for custom contrast
}
\details{
The test can be performed either on raw data on a d coefficient. When \code{d} is provided, also the total sample size \code{n}
         and the scaling method \code{scale} should be provided.

     The parameter \code{scale} controls the method used to scale the effect size d.
     \enumerate{
    \item    \code{scale="g"} assumes scaling by dividing 2*d by the sum of absolute coefficients
    \item    \code{scale="z"} assumes scaling by dividing d by the square-root of the sum of squares of the contrast weigths
    \item    \code{numeric} any constant that multiplies the unscaled d to obtain the scaled d
   }
}
\examples{
cont<-c(-3,-1,1,3)
means<-c(10,12,10,12)
y<-rep(means,1000)+rnorm(4000,0,10)
x<-rep(1:4,1000)
dat<-as.data.frame(cbind(y,x))
dat$x<-factor(dat$x)

test.contr(data = dat,cont=cont,yname = "y",xname = "x")

# check the contrast value
observed<-tapply(dat$y,dat$x,mean)
observed\%*\%cont

# check the t-test and p-value
contrasts(dat$x)<-contr.custom(cont)
summary(lm(y~x,data=dat))
}
\author{
Marcello Gallucci, \email{mcfanda@gmail.com}
}
\keyword{comparisons,}
\keyword{contrasts,}
\keyword{planned}
\keyword{power,}
\keyword{t-test}
